backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# AI & NLP
openai==1.12.0
anthropic==0.18.1
langchain==0.1.0
tiktoken==0.5.2

# Blockchain
web3==6.15.1
eth-account==0.11.0
eth-utils==4.0.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Cache
redis==5.0.1

# WebSocket
python-socketio==5.11.0
websockets==12.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# Utilities
python-dotenv==1.0.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Logging
loguru==0.7.2

